Add a script to control Redmine
For now this is limited to versions.

The close_redmine_version script which does the right thing during the
release process.

There is no create_redmine_version script since that requires some
calculation of the next version and I didn't want to touch that just
yet. The direct redmine script can be used instead.