Support git worktrees for foreman-packaging
Previously it was assumed to have something like ~/dev/foreman-packaging
but with git worktrees you can create multiple working directories. I
have switched over to:

    $ tree -L 1 ~/dev/foreman-packaging
    /home/ekohl/dev/foreman-packaging
    ├── deb-3.1
    ├── deb-3.2
    ├── deb-3.3
    ├── deb-3.4
    ├── deb-3.5
    ├── deb-3.6
    ├── deb-develop
    ├── master
    ├── rpm-3.1
    ├── rpm-3.2
    ├── rpm-3.3
    ├── rpm-3.5
    ├── rpm-3.6
    └── rpm-develop

This is all backed by a single .git directory, which makes it much
cheaper.

It is off by default, but can be enabled with a (local) setting.