Use branching date as target_date in branching procedures
All procedures currently accept a single target_date. For branching
procedures in particular, if the branching date is used as the
target_date, this generates correct dates in the section headings
but an incorrect date in the command invocation to generate the
release procedure. This commit fixes the inconsistency by assuming
the branch date is the target_date in branching, and adding four
weeks to that date to get the date for GA.