Extend PR template after branching docs
In "foreman-documentation", there are a lot of open upstream PRs and
there is typically a lot of cherry-picking to release branches. When we
push the new release branch to foreman-documantation, we should
immediately add this as a friendly reminder to a) users opening new PRs
and b) maintains that cherry-pick PRs to the PR template.