create empty repository when reposync from prod fails
when there is no prod repo, reposync will fail. this seems to be OK on
EL7 and Fedora, but the DNF in EL8 exits non-zero in this case and
this interrupts the stage repository building.
just catching this error is not sufficient, as then repodiff will fail
on the non-existing repository, so we have to create an empty one.