clean up after ContentTooShortError
sometimes urlretrieve will not get the whole file and raise a
ContentTooShortError. it will not delete the partially downloaded file
in this case, which in a later run will cause issues as "the file is
already there" but "it's not a valid RPM file".

instead, catch the ContentTooShortError and try to delete the file