Verify VERSION inside the tarball is correct
Previously I always did this by hand, but it's rather trivial to verify
automatically.

The manual inspection using file-roller is kept in, but arguably there's
less need for that now. In the past I've made sure versions were correct
(mostly applies to the installer), but the point of this check was
always to verify Jenkins built correct tarballs.