Remove PASS_NAME_{GPG,KEY} from main settings file
The values were specific to Foreman, but we have other projects. The
code is now modified to ensure the variables are set before they're
used.

It now also loads the common Foreman settings file for projects that
inherit the Foreman settings. Previously it only loaded the version
specific settings.

Some care is taken in Foreman's common settings to only export values
when no inheritance is taking place.